Commit changes from coding agent for task attempt faf9032c-6400-46f5-a1b8-7fa39dc79f0a
